refactor(graphql.ts): add a graphql module under mina dir
Adds graphql.ts which is a module to hold all response type and graphql
query resources when interacting with the mina daemon graphql. We remove
these from fetch.ts as fetch was starting to get bloated and harder to
read. This aims to be a refactor that offers a cleaner seperation of
concerns between the graphql resources and the actual requests being
sent
