+++ src/index.ts
@@ -72,7 +72,7 @@ export {
   TransactionVersion,
 } from './lib/account_update.js';
 
-export type { TransactionStatus } from './lib/fetch.js';
+export type { TransactionStatus } from './lib/mina/graphql.js';
 export {
   fetchAccount,
   fetchLastBlock,
