+++ src/lib/fetch.ts
@@ -15,6 +15,25 @@ import {
   parseFetchedAccount,
   PartialAccount,
 } from './mina/account.js';
+import {
+  type LastBlockQueryResponse,
+  type GenesisConstants,
+  type LastBlockQueryFailureCheckResponse,
+  type FetchedBlock,
+  type TransactionStatus,
+  type TransactionStatusQueryResponse,
+  type EventQueryResponse,
+  type ActionQueryResponse,
+  type EventActionFilterOptions,
+  type SendZkAppResponse,
+  sendZkappQuery,
+  lastBlockQuery,
+  lastBlockQueryFailureCheck,
+  transactionStatusQuery,
+  getEventsQuery,
+  getActionsQuery,
+  genesisConstantsQuery,
+} from './mina/graphql.js';
 
 export {
   fetchAccount,
@@ -27,9 +46,6 @@ export {
   markActionsToBeFetched,
   fetchMissingData,
   fetchTransactionStatus,
-  TransactionStatus,
-  EventActionFilterOptions,
-  SendZkAppResponse,
   getCachedAccount,
   getCachedNetwork,
   getCachedActions,
@@ -42,13 +58,13 @@ export {
   setArchiveGraphqlEndpoint,
   setArchiveGraphqlFallbackEndpoints,
   setLightnetAccountManagerEndpoint,
-  sendZkappQuery,
   sendZkapp,
   removeJsonQuotes,
   fetchEvents,
   fetchActions,
   Lightnet,
   type GenesisConstants,
+  type ActionStatesStringified,
 };
 
 type NetworkConfig = {
@@ -220,16 +236,6 @@ type FetchError = {
 type ActionStatesStringified = {
   [K in keyof ActionStates]: string;
 };
-type GenesisConstants = {
-  genesisTimestamp: string;
-  coinbase: number;
-  accountCreationFee: number;
-  epochDuration: number;
-  k: number;
-  slotDuration: number;
-  slotsPerEpoch: number;
-};
-
 // Specify 5min as the default timeout
 const defaultTimeout = 5 * 60 * 1000;
 
@@ -451,111 +457,6 @@ async function fetchLastBlock(graphqlEndpoint = networkConfig.minaEndpoint) {
   return network;
 }
 
-type EpochData = {
-  ledger: {
-    hash: string;
-    totalCurrency: string;
-  };
-  seed: string;
-  startCheckpoint: string;
-  lockCheckpoint: string;
-  epochLength: string;
-};
-
-type LastBlockQueryResponse = {
-  bestChain: {
-    protocolState: {
-      blockchainState: {
-        snarkedLedgerHash: string;
-        stagedLedgerHash: string;
-        date: string;
-        utcDate: string;
-        stagedLedgerProofEmitted: boolean;
-      };
-      previousStateHash: string;
-      consensusState: {
-        blockHeight: string;
-        slotSinceGenesis: string;
-        slot: string;
-        nextEpochData: EpochData;
-        stakingEpochData: EpochData;
-        epochCount: string;
-        minWindowDensity: string;
-        totalCurrency: string;
-        epoch: string;
-      };
-    };
-  }[];
-};
-
-const lastBlockQuery = `{
-  bestChain(maxLength: 1) {
-    protocolState {
-      blockchainState {
-        snarkedLedgerHash
-        stagedLedgerHash
-        date
-        utcDate
-        stagedLedgerProofEmitted
-      }
-      previousStateHash
-      consensusState {
-        blockHeight
-        slotSinceGenesis
-        slot
-        nextEpochData {
-          ledger {hash totalCurrency}
-          seed
-          startCheckpoint
-          lockCheckpoint
-          epochLength
-        }
-        stakingEpochData {
-          ledger {hash totalCurrency}
-          seed
-          startCheckpoint
-          lockCheckpoint
-          epochLength
-        }
-        epochCount
-        minWindowDensity
-        totalCurrency
-        epoch
-      }
-    }
-  }
-}`;
-
-type FailureReasonResponse = {
-  failures: string[];
-  index: number;
-}[];
-
-type LastBlockQueryFailureCheckResponse = {
-  bestChain: {
-    transactions: {
-      zkappCommands: {
-        hash: string;
-        failureReason: FailureReasonResponse;
-      }[];
-    };
-  }[];
-};
-
-const lastBlockQueryFailureCheck = (length: number) => `{
-  bestChain(maxLength: ${length}) {
-    transactions {
-      zkappCommands {
-        hash
-        failureReason {
-          failures
-          index
-        }
-      }
-    }
-  }
-}`;
-
 async function fetchLatestBlockZkappStatus(
   blockLength: number,
   graphqlEndpoint = networkConfig.minaEndpoint
@@ -608,48 +509,6 @@ async function checkZkappTransaction(txnId: string, blockLength = 20) {
   };
 }
 
-type FetchedBlock = {
-  protocolState: {
-    blockchainState: {
-      snarkedLedgerHash: string; // hash-like encoding
-      stagedLedgerHash: string; // hash-like encoding
-      date: string; // String(Date.now())
-      utcDate: string; // String(Date.now())
-      stagedLedgerProofEmitted: boolean; // bool
-    };
-    previousStateHash: string; // hash-like encoding
-    consensusState: {
-      blockHeight: string; // String(number)
-      slotSinceGenesis: string; // String(number)
-      slot: string; // String(number)
-      nextEpochData: {
-        ledger: {
-          hash: string; // hash-like encoding
-          totalCurrency: string; // String(number)
-        };
-        seed: string; // hash-like encoding
-        startCheckpoint: string; // hash-like encoding
-        lockCheckpoint: string; // hash-like encoding
-        epochLength: string; // String(number)
-      };
-      stakingEpochData: {
-        ledger: {
-          hash: string; // hash-like encoding
-          totalCurrency: string; // String(number)
-        };
-        seed: string; // hash-like encoding
-        startCheckpoint: string; // hash-like encoding
-        lockCheckpoint: string; // hash-like encoding
-        epochLength: string; // String(number)
-      };
-      epochCount: string; // String(number)
-      minWindowDensity: string; // String(number)
-      totalCurrency: string; // String(number)
-      epoch: string; // String(number)
-    };
-  };
-};
-
 function parseFetchedBlock({
   protocolState: {
     blockchainState: { snarkedLedgerHash, utcDate },
@@ -695,14 +554,6 @@ function parseEpochData({
   };
 }
 
-type TransactionStatusQueryResponse = {
-  transactionStatus: TransactionStatus;
-};
-
-const transactionStatusQuery = (txId: string) => `query {
-  transactionStatus(zkappTransaction:"${txId}")
-}`;
-
 /**
  * Fetches the status of a transaction.
  */
@@ -724,27 +575,6 @@ async function fetchTransactionStatus(
 }
 
 /**
- * INCLUDED: A transaction that is on the longest chain
- *
- * PENDING: A transaction either in the transition frontier or in transaction pool but is not on the longest chain
- *
- * UNKNOWN: The transaction has either been snarked, reached finality through consensus or has been dropped
- *
- */
-type TransactionStatus = 'INCLUDED' | 'PENDING' | 'UNKNOWN';
-
-type SendZkAppResponse = {
-  sendZkapp: {
-    zkapp: {
-      hash: string;
-      id: string;
-      zkappCommand: ZkappCommand;
-      failureReasons: FailureReasonResponse;
-    };
-  };
-};
-
-/**
  * Sends a zkApp command (transaction) to the specified GraphQL endpoint.
  */
 function sendZkapp(
@@ -761,161 +591,6 @@ function sendZkapp(
     }
   );
 }
-
-// TODO: Decide an appropriate response structure.
-function sendZkappQuery(json: string) {
-  return `mutation {
-  sendZkapp(input: {
-    zkappCommand: ${removeJsonQuotes(json)}
-  }) {
-    zkapp {
-      hash
-      id
-      failureReason {
-        failures
-        index
-      }
-      zkappCommand {
-        memo
-        feePayer {
-          body {
-            publicKey
-          }
-        }
-        accountUpdates {
-          body {
-            publicKey
-            useFullCommitment
-            incrementNonce
-          }
-        }
-      }
-    }
-  }
-}
-`;
-}
-type EventQueryResponse = {
-  events: {
-    blockInfo: {
-      distanceFromMaxBlockHeight: number;
-      globalSlotSinceGenesis: number;
-      height: number;
-      stateHash: string;
-      parentHash: string;
-      chainStatus: string;
-    };
-    eventData: {
-      transactionInfo: {
-        hash: string;
-        memo: string;
-        status: string;
-      };
-      data: string[];
-    }[];
-  }[];
-};
-
-type ActionQueryResponse = {
-  actions: {
-    blockInfo: {
-      distanceFromMaxBlockHeight: number;
-    };
-    actionState: {
-      actionStateOne: string;
-      actionStateTwo: string;
-    };
-    actionData: {
-      accountUpdateId: string;
-      data: string[];
-    }[];
-  }[];
-};
-
-type EventActionFilterOptions = {
-  to?: UInt32;
-  from?: UInt32;
-};
-
-const getEventsQuery = (
-  publicKey: string,
-  tokenId: string,
-  filterOptions?: EventActionFilterOptions
-) => {
-  const { to, from } = filterOptions ?? {};
-  let input = `address: "${publicKey}", tokenId: "${tokenId}"`;
-  if (to !== undefined) {
-    input += `, to: ${to}`;
-  }
-  if (from !== undefined) {
-    input += `, from: ${from}`;
-  }
-  return `{
-  events(input: { ${input} }) {
-    blockInfo {
-      distanceFromMaxBlockHeight
-      height
-      globalSlotSinceGenesis
-      stateHash
-      parentHash
-      chainStatus
-    }
-    eventData {
-      transactionInfo {
-        hash
-        memo
-        status
-      }
-      data
-    }
-  }
-}`;
-};
-const getActionsQuery = (
-  publicKey: string,
-  actionStates: ActionStatesStringified,
-  tokenId: string,
-  _filterOptions?: EventActionFilterOptions
-) => {
-  const { fromActionState, endActionState } = actionStates ?? {};
-  let input = `address: "${publicKey}", tokenId: "${tokenId}"`;
-  if (fromActionState !== undefined) {
-    input += `, fromActionState: "${fromActionState}"`;
-  }
-  if (endActionState !== undefined) {
-    input += `, endActionState: "${endActionState}"`;
-  }
-  return `{
-  actions(input: { ${input} }) {
-    blockInfo {
-      distanceFromMaxBlockHeight
-    }
-    actionState {
-      actionStateOne
-      actionStateTwo
-    }
-    actionData {
-      accountUpdateId
-      data
-    }
-  }
-}`;
-};
-const genesisConstantsQuery = `{
-    genesisConstants {
-      genesisTimestamp
-      coinbase
-      accountCreationFee
-    }
-    daemonStatus {
-      consensusConfiguration {
-        epochDuration
-        k
-        slotDuration
-        slotsPerEpoch
-      }
-    }
-  }`;
 
 /**
  * Asynchronously fetches event data for an account from the Mina Archive Node GraphQL API.
